# docs/swagger.yaml
definitions:
  models.Response:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
      status:
        type: string
    type: object
host: localhost:8000
info:
  contact: {}
  description: API Documentation for Spotify
  title: Spotify API
  version: 1.0.0
paths:
  /addFavSong:
    get:
      consumes:
      - application/json
      description: Add your Fav. Song with this api
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Response'
      tags:
      - User
    post:
      consumes:
      - application/json
      description: Add your Fav. Song with this api
      parameters:
      - description: enter song id
        in: body
        name: details
        required: true
        schema:
          example: '{"id":"xyz"}'
          type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Response'
      tags:
      - User
...
  /addToRecentlyPlayed:
    post:
      consumes:
      - application/json
      description: Add to Recently_Played list
      parameters:
      - description: enter song id
        in: body
        name: details
        required: true
        schema:
          example: '{"song_id":"xyz"}'
          type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Response'
      tags:
      - User
...
  /createPlaylist:
    post:
      consumes:
      - application/json
      description: Create playlist
      parameters:
      - description: enter playlist name and song id
        in: body
        name: details
        required: true
        schema:
          example: '{"playlist_name":"name of your playlist","song_id":"xyz"}'
          type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Response'
      tags:
      - User
  /getAlbum:
    post:
      consumes:
      - application/json
      description: Get Album
      parameters:
      - description: enter name of album SchemaExample({
        in: body
        name: details
        required: true
        schema:
          example: '{"album_name":"xyzBest"}'
          type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Response'
      tags:
      - User
...
  /getArtist:
    post:
      consumes:
      - application/json
      description: Get Artist
      parameters:
      - description: enter name of artist SchemaExample({
        in: body
        name: details
        required: true
        schema:
          example: '{"artist_name":"Arijit Singh"}'
          type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Response'
      tags:
      - User
...
  /getSong:
    post:
      consumes:
      - application/json
      description: Get song by id
      parameters:
      - description: enter song id
        in: body
        name: details
        required: true
        schema:
          example: '{"song_id":"xyz"}'
          type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Response'
      tags:
      - User
...
  /searchSongs:
    post:
      consumes:
      - application/json
      description: Search Song by name
      parameters:
      - description: enter name of song SchemaExample({
        in: body
        name: details
        required: true
        schema:
          example: '{"name":"song_name"}'
          type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Response'
      tags:
      - User
...
  /userLogin:
    post:
      consumes:
      - application/json
      description: User login with name and contact_no
      parameters:
      - description: name and contact number
        in: body
        name: details
        required: true
        schema:
          example: '{"name":"john doe","contact":"1234567890"}'
          type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Response'
      tags:
      - User
...
swagger: "2.0"